jwt verification, access and refresh tokens

+++ src/utils/cloudinary.js
@@ -1,25 +1,47 @@
-import { v2 as cloudinary } from "cloudinary";
-import fs from "fs";
+import {v2 as cloudinary} from "cloudinary"
+import fs from "fs"
+
+console.log("Before config:");
+console.log(process.env.CLOUDINARY_CLOUD_NAME);
+console.log(process.env.CLOUDINARY_API_KEY);
+console.log(process.env.CLOUDINARY_API_SECRET);
 
 cloudinary.config({ 
-    cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
-    api_key: process.env.CLOUDINARY_API_KEY, 
-    api_secret: process.env.CLOUDINARY_API_SECRET 
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+  api_key: process.env.CLOUDINARY_API_KEY, 
+  api_secret: process.env.CLOUDINARY_API_SECRET 
 });
+
+console.log("Cloudinary Config:");
+console.log(cloudinary.config());
 
 const uploadOnCloudinary = async (localFilePath) => {
     try {
         if (!localFilePath) return null
+
+        console.log("Uploading:", localFilePath);
+console.log("Cloud Name:", process.env.CLOUDINARY_CLOUD_NAME);
         //upload the file on cloudinary
         const response = await cloudinary.uploader.upload(localFilePath, {
             resource_type: "auto"
         })
         // file has been uploaded successfull
-        console.log("file is uploaded on cloudinary ", response.url);
+        //console.log("file is uploaded on cloudinary ", response.url);
+        fs.unlinkSync(localFilePath)
         return response;
-        
+
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
-        return null;
+    console.log("========== CLOUDINARY ERROR ==========");
+    console.error(error);
+
+    if (localFilePath && fs.existsSync(localFilePath)) {
+        fs.unlinkSync(localFilePath);
     }
+
+    return null;
 }
+}
+
+
+
+export {uploadOnCloudinary}
